requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
pyaml==19.4.1
PyYAML==5.1.2
requests==2.22.0
urllib3==1.25.7
